+++ module/sheets/actor-sheet.mjs
@@ -1,7 +1,7 @@
 import { onManageActiveEffect, prepareActiveEffectCategories } from "../helpers/effects.mjs";
-import { getItemsOfType, searchCompendium } from "../helpers/utils.mjs";
+import { getItemsOfType, rememberOptions, searchCompendium } from "../helpers/utils.mjs";
 import { onMorph, onZordDelete, prepareZords } from "./power-ranger-sheet-helper.mjs";
-// import {  } from "./transformer-sheet-helper.mjs";
+import { onAltModeDelete, onTransform } from "./transformer-sheet-helper.mjs";
 
 /**
  * Extend the basic ActorSheet with some very simple modifications
@@ -116,7 +116,7 @@ export class Essence20ActorSheet extends ActorSheet {
         buttons: {
           save: {
             label: game.i18n.localize('E20.AcceptButton'),
-            callback: html => this._crossoverSettings(this._rememberOptions(html)),
+            callback: html => this._crossoverSettings(rememberOptions(html)),
           },
         },
       },
@@ -343,7 +343,7 @@ export class Essence20ActorSheet extends ActorSheet {
       if (item.type == "origin") {
         this._onOriginDelete(item);
       } else if (item.type == "altMode") {
-        this._onAltModeDelete(item);
+        onAltModeDelete(item, this);
       } else if (item.type == 'influence') {
         this._onInfluenceDelete(item);
       }
@@ -365,7 +365,7 @@ export class Essence20ActorSheet extends ActorSheet {
     html.find('.morph').click(() => onMorph(this));
 
     //Transform Button
-    html.find('.transform').click(this._transform.bind(this));
+    html.find('.transform').click(() => onTransform(this));
 
     // Rollable abilities.
     if (this.actor.isOwner) {
@@ -425,151 +425,6 @@ export class Essence20ActorSheet extends ActorSheet {
         li.setAttribute("draggable", true);
         li.addEventListener("dragstart", handler, false);
       });
-    }
-  }
-
-
-  /**
-  * Handle clicking the transform button
-  * @private
-  */
-  async _transform() {
-    const altModes = getItemsOfType("altMode", this.actor.items);
-
-    if (!altModes.length && !this.actor.system.isTransformed) { // No alt-modes to transform into
-      ui.notifications.warn(game.i18n.localize('E20.AltModeNone'));
-    } else if (altModes.length > 1) {                           // Select from multiple alt-modes
-      if (!this.actor.system.isTransformed) {
-        this._showAltModeChoiceDialog(altModes, false);         // More than 1 altMode and not transformed
-      } else {
-        this._showAltModeChoiceDialog(altModes, true);          // More than 1 altMode and transformed
-      }
-    } else {                                                    // Alt-mode/bot-mode toggle
-      this.actor.system.isTransformed ? this._transformBotMode() : this._transformAltMode(altModes[0]);
-    }
-  }
-
-  /**
-   * Creates the Alt Mode Choice List Dialog
-   * @param {AltMode[]} altModes  A list of the available Alt Modes
-   * @param {Boolean} isTransformed Whether the Transformer is transformed or not
-   * @private
-   */
-  async _showAltModeChoiceDialog(altModes, isTransformed) {
-    const choices = {};
-    if (isTransformed) {
-      choices["BotMode"] = {
-        chosen: false,
-        label: "BotMode",
-      };
-    }
-
-    for (const altMode of altModes) {
-      if (this.actor.system.altModeId != altMode._id) {
-        choices[altMode._id] = {
-          chosen: false,
-          label: altMode.name,
-        };
-      }
-    }
-
-    new Dialog(
-      {
-        title: game.i18n.localize('E20.AltModeChoice'),
-        content: await renderTemplate("systems/essence20/templates/dialog/option-select.hbs", {
-          choices,
-        }),
-        buttons: {
-          save: {
-            label: game.i18n.localize('E20.AcceptButton'),
-            callback: html => this._altModeSelect(altModes, this._rememberOptions(html)),
-          },
-        },
-      },
-    ).render(true);
-  }
-
-  /**
-   * Handle selecting an alt-mode from the Alt-mode Dialog
-   * @param {AltMode[]} altModes  A list of the available Alt Modes
-   * @param {Object} options   The options resulting from _showAltModeDialog()
-   * @private
-   */
-  async _altModeSelect(altModes, options) {
-    let selectedForm = null;
-    let transformation = null;
-
-    for (const [altMode, isSelected] of Object.entries(options)) {
-      if (isSelected) {
-        selectedForm = altMode;
-        break;
-      }
-    }
-
-    if (!selectedForm) {
-      return;
-    }
-
-    if (selectedForm == "BotMode") {
-      this._transformBotMode();
-    } else {
-      for (const mode of altModes) {
-        if (selectedForm == mode._id) {
-          transformation = mode;
-          break;
-        }
-      }
-
-      if (transformation) {
-        this._transformAltMode(transformation);
-      }
-    }
-  }
-
-  /**
-   * Handles Transforming into an altMode
-   * @param {AltMode} altMode   The alt-mode that was selected to Transform into
-   * @private
-   */
-  async _transformAltMode(altMode) {
-    await this.actor.update({
-      "system.movement.aerial.altMode": altMode.system.altModeMovement.aerial,
-      "system.movement.swim.altMode": altMode.system.altModeMovement.aquatic,
-      "system.movement.ground.altMode": altMode.system.altModeMovement.ground,
-      "system.altModeSize": altMode.system.altModesize,
-      "system.altModeId": altMode._id,
-      "system.isTransformed": true,
-    }).then(this.render(false));
-  }
-
-  /**
-   * Handle Transforming back into the Bot Mode
-   * @private
-   */
-  async _transformBotMode() {
-    await this.actor.update({
-      "system.movement.aerial.altMode": 0,
-      "system.movement.swim.altMode": 0,
-      "system.movement.ground.altMode": 0,
-      "system.isTransformed": false,
-      "system.altModeId": "",
-      "system.altModeSize": "",
-    }).then(this.render(false));
-  }
-
-  /**
-   * Handle AltModes being deleted
-   * @param {AltMode} altMode is the deleted AltMode.
-   * @private
-   */
-  async _onAltModeDelete(altMode) {
-    const altModes = getItemsOfType("altMode", this.actor.items);
-    if (altModes.length > 1) {
-      if (altMode._id == this.actor.system.altModeId) {
-        this._transformBotMode();
-      }
-    } else {
-      this._transformBotMode();
     }
   }
 
@@ -774,7 +629,7 @@ export class Essence20ActorSheet extends ActorSheet {
         buttons: {
           save: {
             label: game.i18n.localize('E20.AcceptButton'),
-            callback: html => this._hangUpSelect(itemArray, this._rememberOptions(html), perkIds, newInfluence),
+            callback: html => this._hangUpSelect(itemArray, rememberOptions(html), perkIds, newInfluence),
           },
         },
       },
@@ -859,26 +714,14 @@ export class Essence20ActorSheet extends ActorSheet {
         buttons: {
           save: {
             label: game.i18n.localize('E20.AcceptButton'),
-            callback: html => this._showOriginSkillDialog(origin, this._rememberOptions(html), event, data),
+            callback: html => this._showOriginSkillDialog(origin, rememberOptions(html), event, data),
           },
         },
       },
     ).render(true);
   }
 
-  /**
-   * Returns values of inputs upon dialog submission. Used for passing data between sequential dialogs.
-   * @param {HTML} html   The html of the dialog upon submission
-   * @returns {Object>}  The dialog inputs and their submitted values
-   * @private
-   */
-  _rememberOptions(html) {
-    const options = {};
-    html.find("input").each((i, el) => {
-      options[el.id] = el.checked;
-    });
-    return options;
-  }
+
 
   /**
    * Displays a dialog for selecting a Skill for the given Origin.
@@ -932,7 +775,7 @@ export class Essence20ActorSheet extends ActorSheet {
         buttons: {
           save: {
             label: game.i18n.localize('E20.AcceptButton'),
-            callback: html => this._originStatUpdate(origin, selectedEssence, this._rememberOptions(html), event, data),
+            callback: html => this._originStatUpdate(origin, selectedEssence, rememberOptions(html), event, data),
           },
         },
       },
